refactor: 테스트 클래스 패키지를 도메인과 동일하게 변경
