+++ src/main/java/ladder/domain/LadderGame.java
@@ -70,3 +70,5 @@ public class LadderGame {
         return ladder.getRewards();
     }
 }
+
+
