+++ src/test/java/ladder/domain/ladder/NoncontinuousRandomLineStrategyTest.java
@@ -1,7 +1,5 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
-import ladder.domain.ladder.NoncontinuousRandomLineStrategy;
-import ladder.domain.ladder.Step;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
