+++ src/test/java/ladder/domain/ladder/RewardsTest.java
@@ -1,6 +1,6 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
-import ladder.domain.ladder.Rewards;
+import ladder.domain.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
